Add a .tox/requirements venv to simplify Makefile
This simplifies the `Makefile` code for compiling requirements files
because every requirements file is now compiled using a tox environment
with a matching name.

Previously this was true of all the requirements files except
`requirements/requirements.txt` which was compiled in a tox env named
`dev` not `requirements`:

    tox -e dev --run-command 'pip-compile requirements/requirements.txt'

Now the tox environment and requirements files names match:

    tox -e requirements --run-command 'pip-compile requirements/requirements.txt'

This allows the `Makefile` code for generating these commands to be
simpler.