Fix mocked async method throwing instead of rejecting
A SubmitGradeForm test was triggering a grade submission, which involves async
steps, and then checking synchronously for the error dialog to be shown. This
should not have worked, since the error dialog can only be shown asynchronously
after the submission completes. However the test made the mock throw
synchronously rather than reject asynchronously.