Raise an exception for disabled accounts
This is a pretty blunt instrument, but any action that attempts to get
the "current" application instance will fail with an "AccountDisabled"
error.

As this is a serialisable error, and serialisable errors are handled in
both the UI and API, this should result in some sensible output in all
cases.