Validate input for search queries
Just validate that the types match what is expected by the services

For integers we need to create a custom field that allows empty strings
to be handled as `None`.

On top of that, if that integer is going to be matched against a
postgres `integer` primary key we need to validate the range to avoid
overflowing.