New `list_all_files` method in BB API client
To populate the file picker in blackboard we make one API request
per folder as the user navigates the tree. This is sensible as it keeps
the API requests themselves small and fast.

Unfortunately for course copy we need the full tree of files to be able
to:

- Find the current file that triggered the course copy process
- Guarantee that subsequent launches of other assignments in the course
  will also find their files without instructor intervention.

We'll use the new method only while doing course copy and keep the
original `list_files` as the driver of the file picker.