Fix the definition of billable to include at least one annotation in the group
It appears we inadvertently changed the definition of billable users during
the creation of report by missing out the requirement that a billable user be
in a group with at least one annotation.

This leads to a larger amount of users being counted as billable as a teacher
can end up including users into other periods of time simply by visiting groups.