Rename LMSGrader => GradingToolbar and convert to TS
The "LMS" part of the component name was redundant, and the "Toolbar" suffix
makes it more obvious what kind of control this is.

This might not be the final name, as we're considering adding an "edit
assignment" button to this toolbar in future and displaying it even when grading
is not active.