Begin to clarify which frontend config is available in different modes
It was unclear which fields in the `ConfigObject` type are present in which
apps, and the optional-ness of many fields did not match the code which used
them. ie. We had code testing for the presence of fields that were marked
non-optional in the type.

This commit begins to clean that up by:

 - Organizing fields in `ConfigObject` by the application modes they
   appear in

 - Enabling the @typescript-eslint/no-unnecessary-condition rule to
   check for conditionals that always evaluate to the truthy or falsey
   according to the types. This helps catch cases where a field is
   incorrectly marked as non-optional.

 - Correcting the errors found by the above lint rule, which included
   incorrect optional-ness for several fields as well as some other
   unnecessary conditionals.

 - Correcting a Canvas SpeedGrader test whose expectations did not match the
   backend code

There are still some inconsistencies in the `ConfigObject` type. Some fields
that are present only in certain modes are still marked as non-optional. However
it is now at least consistent with the code that uses it. This will be further
cleaned up in future.