Add AssignmentService.get_copied_from_assignment
This makes the behaviour of DocumentUrlService more explicit around
course copy.

Ideally the final state of this would be to have each of the possible
parameter locations be part of the individual product/product plugins instead of
trying them in succession. Unfortunately the exact behaviour we rely on
doesn't seem to be documented in D2L / Blackboard and we might be
relying in it for other smaller LMS's even if it's not documented here.

Keeping the same behaviour for now in all LMS's while making the course
copy aspect of it more apparent.