Make AssignmentService.upsert_assignment aware of course copy
Use LTIParams in combination with the new `get_copied_from_assignment`
to, in new assignments, copy some of the configuration of the old
assignment over to he new one.

We only copy the value of `group_set_id` for now. As a result copied
assignments will still be group ones after copy. They will point to an
unavailable group set for now.

We don't look at document_url now as that handled separately at the
moment over in DocumentService.