Record course in the DB on configuration
Create a course in the DB as soon as we start configuring an assignment
in it.

This brings the behaviour in line with the deeplinking counterpart where
this was already the case, the deeplinking request creates the course.

The new behaviour allows APIs that are called during
configuration (files, groups) to rely on the presence of the course.