Use a single name for searching for instances which matches the pattern
Previously we had `admin.instances` to start the search and `admin.instances.search`
for the callback. In all other cases we have the same route and distinguish by
HTTP method. This keeps the two halves of the search more obviously intertwined.

We have `admin.instance.blah` for pretty much everything else, so I've made this
`admin.instance.search` and renamed the template accordingly.