Add `raise_on_missing` to CourseService.get_by_context_id
This parameter/flag allows to clearly express the expected value
from the method and use the right SQLA method accordingly.

Updating the callers that expect a course in all cases to, in case of
a bug, get a more clear trace instead of the first attribute access
failing AttributeError on NoneType.