PoC of editing assignments with server-side transition from View => Edit
Add a proof of concept of initiating an assignment edit via a server-side
transition when clicking "Edit assignment" in the instructor toolbar. This works
by having the frontend create and submit a hidden form when clicking "Edit
assignment".

The upside is that this code is very simple. The principle downside is that
clicking the "Back" button shows an error message in Chrome [1] warning that
loading the page will involve a form-resubmission.

[1] https://hypothes-is.slack.com/archives/C4K6M7P5E/p1677243443463259?thread_ts=1677174606.567559&cid=C4K6M7P5E