Handle submissions with a limited number of attempts
This seems to be only handled by the LTI1.3 version of the API on the
Canvas side.

For now we are just swallowing the error. We'll instruct teachers to
not limit the number of submission for H assignments.

If we eventually want to display some useful information to the user we
can attach an error code here.