Support pre-selecting content in ContentSelector
Add `initialContent` prop which specifies an existing configuration, and use the
value to pre-fill the `URLPicker` and `JSTORPicker` inputs when selected by the
user. The content picker that corresponds to `initialContent` is not
automatically opened because that would make the action of the "Cancel" button
in the picker unclear - does it close the picker or exit the content selection
screen entirely?