Factor out common ErrorModal props in LaunchErrorDialog
This will make it easier to add additional props to all `ErrorModal` cases in
future. For the `onRetry` prop, I defaulted to enabling this action, but
retained the existing behavior for the many cases which disable it and added a
note that we might want to revisit this in future.