Redesign grade validation and formatting
Rewrite the grade formatting/validation functions to create a clearer separation
of responsibilities between them and the `SubmitGradeForm` component, and handle
validation errors in a clearer way. The revised code has two functions:

 - `formatGrade` converts a value received from the API into a string
   for the grade input field.

 - `validateGrade` converts user input into either a grade value for
   submission to the API, or a validation error. This replaces/combines
   the `formatToNumber` and `validateGrade` functions that existed
   previously.

The handling of grades that don't map to integer values has also been
simplified. We now display a maximum of 2 digits after the decimal point,
which seems adequate for all reasonable usage.