Query the DB directly for files that belong to the course
This is not only more explicit, just query what we are after, but also
fixes a bug where the same file ID might exist in different courses
raising a MultipleResultException.

With this version we'll only get the file if that's the one we are
looking for returning None/False in any other case.