Add a new, rewritten `DigestService`
The main aim of this rewrite is to pave the way for replacing lots of
little DB queries (with the number of queries growing based on the input
data) with a fixed number of larger queries.

`DigestContext.unified_users()` and `DigestContext.unified_courses()`
are methods that do lots of DB queries but the queries are now only run
once per `DigestContext` instance. This means that the bodies of these
methods can now be refactored to replace lots of little queries with a
fixed number of larger ones.