Disable auto-submission after selecting content if editing existing assignment
When configuring a new assignment, the form auto-submits after choosing the
content if there is no other configuration available. When editing an existing
assignment, it is simpler / makes more sense if the flow is always the same,
regardless of whether groups are available to configure or not. We might want to
disable auto-submission in all cases in future, to reduce the number of
scenarios we have to test.