Enable assignment editing for all institutions
Show the editing controls for all teachers in non-Canvas launches.

Doing this as a code change instead of a migration as it simplifies the
code a bit and also allows an hypothetical rollback (a revert in this c
ase) to go back to the exact same state relying on the DB data.