Handle Mailchimp API errors when sending email digests
Fixes https://github.com/hypothesis/lms/issues/5276

**Problem:** `DigestService` loops over a batch of users calling
`MailchimpService.send_template()` for each user. Currently there's no
handling of any exceptions raised by `MailchimpService.send_template()`
so if trying to email one user raises an exception then it'll crash out
of the `for` loop and none of the following users in the batch will get
their digest emails either.

**Solution:** catch any exceptions from `MailchimpService` and continue
through the `for` loop, but also log the exception and report it to
Sentry so that we don't swallow Mailchimp errors silently.

After the `for` loop we still raise a final exception if any of the
iterations failed so that the calling code can know that something went
wrong. This should enable the Celery task to trigger a retry.