Fix plain SQL execution on SQLA
Fixes:
sqlalchemy.exc.RemovedIn20Warning:

The Engine.execute() method is considered legacy as of the 1.x series of SQLAlchemy
and will be removed in 2.0. All statement execution in SQLAlchemy 2.0 is performed
by the Connection.execute() method of Connection, or in the ORM by the Session.execute() method of Session.
(Background on SQL Alchemy 2.0 at: https://sqlalche.me/e/b8d9)