Add a default renderer instead of using override_renderer
Without the default we get:

```
ValueError: Could not convert return value of the view callable function
lms.views.email.unsubscribe into a response object. The value returned was {}.
You may have forgotten to define a renderer in the view configuration.
```

even if the template does indeed render correctly but we get a 500
error.

With this change we get the same response but with a 200 instead.