Use a durable RabbitMQ exchange and queue
Use a durable RabbitMQ exchange and queue so that any messages are
persisted to disk and will survive after a RabbitMQ crash or restart.

A durable exchange and queue may not have been necessary when all we had
was the `rotate_keys()` periodic task (although it wouldn't have done
any harm) but now that we have the email digests tasks we want our
messages to survive RabbitMQ crashes or restarts. We don't want queued
messages to be lost and emails not to be delivered.

In general we should set `durable=True` unless we have a good reason not
to: it generally won't do any harm and often it will be important not to
lose messages in RabbitMQ restarts.

In order for messages to actually be persisted to disk the
`delivery_mode` of the message also needs to be set to `2`. Celery does
this by default and I've tested locally that persistent delivery mode is
being used for LMS's messages.

For docs on RabbitMQ **Message durability** see here:
https://www.rabbitmq.com/tutorials/tutorial-two-python.html

Celery uses persistent delivery mode by default:
https://docs.celeryq.dev/en/stable/userguide/configuration.html?highlight=persistent#task-default-delivery-mode