Remove unnecessary manual Celery retries
This effectively reverts https://github.com/hypothesis/lms/pull/5279 and
https://github.com/hypothesis/lms/pull/5281 and replaces them with
automatic Celery retries.

The purpose of the manual retries was to retry a task but only for the
sub-tasks that had failed. For example retry the
`send_instructor_email_digests()` but only try to re-send those emails
that failed to send the last time, don't send any duplicates of emails
that have already sent successfully.

Now that we have the `task_done` table
https://github.com/hypothesis/lms/pull/5347 it's no longer necessary to
do this: `MailchimpService.send_template()` won't re-send any emails
that it's previously sent and recorded in the DB.

The downside of this commit is that we may do some unnecessary
processing in the case of a retry to re-generate the data for emails
that we've already sent (even though `send_template()` won't actually
send any duplicates). But we probably don't need to worry about this too
much: retries should be rare so we don't need to optimise for them, and
Celery's exponential backoff and jitter will spread the retries over
time. The benefit of this commit is that by removing manual retries the
code and tests are much simpler.