Remove unnecessary send_instructor_email_digests() error handling
`DigestService.send_instructor_email_digests()` is careful to collect
any errors raised when calling `send_template.delay()` into a
`SendDigestsError` and raise it, but as of
https://github.com/hypothesis/lms/pull/5351 the
`send_instructor_email_digests()` task (which calls
`DigestService.send_instructor_email_digests()`) no longer catches these
`SendDigestsError`'s, so simplify `DigestService` by removing this error
handling.

This does mean that if any one of `DigestService`'s calls to
`send_template.delay()` raises then the whole task will fail and break
out of the `for` loop skipping any remaining calls to
`send_template.delay()` that were still to be made. But Celery will
retry the task which will re-start the `for` loop from the start and
hopefully on the second or third attempt it will make it all the way
through the loop.

This behaviour of catching exceptions and continuing through the `for`
loop made more sense when `DigestService` used to call the Mailchimp API
directly in the `for` loop but seems less necessary now that it the
`for` loop just schedules instances of `send_template.delay()` (which do
the Mailchimp API calls in a separate task): RabbitMQ seems more
reliable than the third-party Mailchimp API.