Add `reject_on_worker_lost` to email tasks
Add `reject_on_worker_lost=True` to the Celery configs of the
email-related tasks.

Apparently by default acknowledges tasks (removes them from the
queue) if the worker process abruptly exits, is terminated, hits a
Python `sys.exit()`, or receives a `KILL` or
`INT` signal. Setting `reject_on_worker_lost=True` makes Celery re-queue
the task instead.

There's good reason why this setting is disabled by default: Celery
doesn't want to re-run a task if the task's code causes a kernel
segfault or similar, or causes the machine to run out of memory, or if a
sysadmin deliberately kills a worker process that's executing the task.
Celery also doesn't want to cause message loops by re-queueing a task
that always fails. The best discussion of it is near the top of this
page in the Celery docs:
https://docs.celeryq.dev/en/stable/userguide/tasks.html
For this reason I don't think we want to enable this setting globally.
Most of our tasks don't need this. But I thought it worth enabling for
the email tasks.

See also:

* https://docs.celeryq.dev/en/stable/reference/celery.app.task.html#celery.app.task.Task.reject_on_worker_lost
* https://docs.celeryq.dev/en/stable/userguide/configuration.html?highlight=task_reject_on_worker_lost#std-setting-task_reject_on_worker_lost