Send the assignment configuration as custom parameters
Custom parameters in the deep message must be present in the launch
payload according to the spec.

This is a cleaner approach than encoding them as query parameters but
unfortunately supported in D2L but not Canvas.

Sending them in all cases as, although canvas doesn't honor the spec
and send them back, it allows them in the message.