Refactor basic launch functional tests
These test were very focused on a feature we don't support, using the
DB as the source of truth for deep linked assignments.

That's complicated to do due to a number of bugs in canvas itself and
we have abandoned that approach. Removed the test covering those.

At the point of the launch the document_url always comes from the URL in
canvas. We should however create an assignment with the right document
url as a side effect of the launch. Asserting that now.