Get files GUID of the current application instance
We store files by application instance so the same file will have one
row per AI for example in the case of LTI upgrades.

We use GUID as a "source of truth" in other tables like assignment and
in general for any communication with the H service (generating group
and user IDs).