Enable clipboard-write permission for assignment content frame
This enables the "Copy transcript" button to work in Via's video player
app. Without this, the button works in a standalone Via tab, but fails
when embedded in the LMS app with this error:

```
The Clipboard API has been blocked because of a permissions policy applied to
the current document. See https://goo.gl/EuHzyv for more details.
```

Note this fix will likely not work if our LMS app is being displayed inside an
iframe instead of being loaded in a new tab, unless the LMS has added the same
permission when it rendered the iframe containing our LMS app.