Enable the YouTube feature by default
Enable the `"youtube.enabled"` feature by default for all application
instances that don't have this feature explicitly set to `false` in the
DB.

This will enable the feature for many application instances but not all:
whenever you edit an application instance using the admin pages it
writes the default value for settings such as `youtube.enabled` into
that application instance's settings in the DB even if that setting
wasn't changed. So application instances that've been edited in admin
pages since the `youtube.enabled` setting was added to the admin pages
will have `false` in the DB for that setting. We'll need to do a DB
migration to change those to `true`.