Add a migration to enable the YouTube feature
Add a DB migration to set the `youtube.enabled` application instance
setting to `true` for all those application instances that currently
have it set to `false`.

Combined with https://github.com/hypothesis/lms/pull/5598 (which enables
`youtube.enabled` for all instances that don't have a value for this
setting in the DB) this will enable the YouTube feature for all
application instances.

This migration is necessary because when you edit an instance in the
admin pages it writes the default value for every setting into the DB
for that instance, even if that setting wasn't edited. So any instances
that've been edited since the `youtube.enabled` setting was added to the
admin pages will have the previous default value of `false` in the DB.