Fix handling of implicit form submission in FilePickerApp
With the addition of the "Title" field to the form, the user can now trigger an
implicit form submission by pressing Enter when that field is focused. When this
happens we need to make sure the same async steps are executed to update the
form before the real submission. This is done by checking for the `shouldSubmit`
flag in `onSubmit`, which indicates that the real submission should happen.

I also added some notes about how we might be able to simplify the whole
form submission process using `requestSubmit` in future.