Handle invalid file URLs explicitly
The URL value is provided by the backend and then forwarded by the
frontend. A URL we can't parse found here would be a bug and hence an
exception we'd like to see risen.

Added explicitly ValueError to get a better message instead of trying to
access any property of None.