Fix `static_url()` in Celery tasks
Fix Pyramid's `static_url()` method when called by Celery tasks. It
currently returns URLs with the wrong hostname and port e.g.
`http://localhost/foo` instead of `http://localhost:8001/foo`.

Fix this by injecting the correct `HTTP_HOST` variable into the WSGI
environment in Celery processes. This is the same WSGI environment that
would normally be set based on the Host header of the HTTP request when
responding to an HTTP request rather than running a Celery task.

I'm not sure if this is the best/right way to do this but it works.

This requires each deployment of the app (dev, QA and prod) to have an
`HTTP_HOST` envvar set correctly.

This replaces a previous solution where we had a
`WEB_APP_URL`/`web_app_url` envvar/setting and we passed a wrapped
`route_url()` method to `EmailUnsubscribeService`:

    route_url=partial(
        request.route_url, _app_url=request.registry.settings["web_app_url"]
    )

This only worked for `route_url` and only for `EmailUnsubscribeService`.
The solution in this commit should fix both `route_url()` and
`static_url()` and should fix them for all code called by Celery tasks.

The previous `WEB_APP_URL` envvar will have to be replaced by the new
`HTTP_HOST` envvar in all environments.