Add `mutate` API to `useFetch` result
This is useful to replace the `useFetch` result, eg. following a local
API call that updates it on the backend, without the delay that actually
re-fetching the value would incur.

As with the `useFetch` API in general, this roughly follows the API of SWR
(https://swr.vercel.app/docs/mutation).