Fix the CSS in the digest email template
For some reason this CSS isn't working in Gmail: Gmail strips the entire
`<style>` tag.

I copy-pasted the CSS from the original Mandrill template again, ran it
through an online CSS formatter (since the formatting of the template in
Mandrill is terrible), and otherwise didn't make any other changes to
the CSS. It now seems to work in Gmail.

If I even indent the CSS so that it's indented more than the containing
`<style>` tag that seems to cause the CSS to break in Gmail.

https://hypothes-is.slack.com/archives/C4K6M7P5E/p1693923095246269