Emit an error code for AIs with provisioning disabled
Just few AI have this disabled and we've seen some launches from them.

This launches completely fail and result in a blank screen.

This commit doesn't remove the concept from the DB to keep track of
which instances have it disabled in case we want to re-enable this.

Alternatively, if no costumers contact us about this we can completely
remove the concept from the DB.