Enable grouped updates for frontend dependencies
This will reduce the overall volume of PRs and conflicts between PRs for
updating frontend dependencies. The configuration was copied from the client
repo.

We don't currently use Sentry in the frontend in this repo, but I left the
configuration on there because I expect we will use it in future.