Add stylesheet for Canvas Pages assignments
Add a stylesheet for the HTML content of Canvas Pages assignments. For reasons
outlined at the top of canvas_pages.scss, this commit uses a re-implementation
of the styling that Canvas provides for Pages. This includes the Lato font,
imported via an Adobe TypeKit project [1] and styling for all of the elements
that can be created using Pages' WYSIWYG editor.

Canvas Pages re-use "common" stylesheets that include > 500KB of styles for
other parts of the UI, which in principle Canvas Pages HTML could reference,
although we assume that most pages do not.

Fixes https://github.com/hypothesis/lms/issues/5746

[1] https://fonts.adobe.com/fonts/lato