Add GUID to LMS annotation metadata
Add the assignment's `tool_consumer_instance_guid` to the metadata that
the LMS app passes to the client for storage with annotations in h.

Also rename `assignment_id` to `resource_link_id` in the metadata so
that it matches the name used for this field in the LMS app's code and
DB, and restructure the metadata so that `resource_link_id` and `guid`
are grouped together in an `assignment` sub-dict.