Fallback to custom params for deep linking configuration in Canvas
In canvas we have historically send the deep linking (DL) configuration
parameters as query parameters in the return URL.

When we started using DL in D2L we instead used LTI custom params which
can be arbitrary JSON which is more ergonomic than having to worry
about encoding issues in URLs.

Before this commit, and as part of the D2L work, we always send the
configuration as custom parameters even if never read them in Canvas,
relying only the URL.

Now, as part of the work with VitalSource we found that URL parameters
are not proxied in their LTI integration.

With this commit we continue reading the parameters from the URL as the
first option, for all existing Canvas installations, but we fallback to
custom parameters, for schools using the VS proxying.