Increase send_instructor_email_digests rate limit
Increase the rate limit of the `send_instructor_email_digests()` Celery
task. This is because I want to try decreasing the task's `batch_size`
and correspondingly increasing its `rate_limit` to see what effect that
has on performance, with an eye towards potentially removing the
batching feature altogether in order to simplify adding future features
to this code.