Rename EmailUnsubscribeService to EmailPreferencesService
This is because we're getting to be adding a `preferences_url()` method
and other public methods to this service, it's not going to be just for
unsubscribes anymore. Also, the service will be changed to read and
write the new `EmailPreferences` model and the current
`EmailUnsubscribe` model will eventually be removed.