Reduce `send_instructor_email_digests` rate limit
Reduce the rate-limit of the `send_instructor_email_digests()` task from `30/m` to `10/m`. This is a desperate attempt to prevent problems that we've seen this morning with a small number of instances of this task failing. We increased the task's throughput yesterday and this morning it started failing. We've seen no evidence that the higher throughput is the cause for the failures, but we don't have any better ideas so let's try reducing the throughput.

Slack thread: https://hypothes-is.slack.com/archives/C0LUWQQJJ/p1700644122131399