Explicitly list all fields send to be stored in group info
Instead of using a dynamic list of fields from GroupInfo.columns
list all possible columns.

These are duplicated information persisted in other tables anyway
but this change makes removing some of them easier to follow in follow
up commits.

This is the last usage of BaseClass.columns() outside the class
itself.

The other usage, in BaseClass itself, is for the override of `__repr__`
implementation. This implementation prints all fields of a
model which often makes it impractical.