Set end CFI when choosing chapter from list if page ranges feature enabled
Previously when the page ranges feature was enabled we would configure the
assignment with either a page range, if the user entered a page range, or a CFI
start point (not a range) if the user chose a chapter. Now when the user chooses
a chapter, we will instead configure the assignment with a CFI range, where the
start point is the CFI of the entry the user selected, and the end point is the
CFI of the next TOC entry at the same level as the one the user selected.

This makes the experience consistent regardless of which selection method the
user chooses.