Fix the email preferences page in Safari
The `/email/preferences?token=xyz` URL currently redirects to
`/email/preferences` (no token) with a `Set-Cookie` header in the HTTP
redirect response.

Unfortunately Safari fails to set cookies when the `Set-Cookie` header
is in a redirect response, meaning email preferences links are currently
broken in Safari.

Fix this by *not* redirecting the browser, and showing the email
preferences UI at the `/email/preferences?token=xyz` URL (*with* token).

The downside of this is that the sensitive token is left in the
browser's location bar, but we plan to have some JavaScript code remove
the token from the URL in a follow-up commit.